+++ app/components/Forms/RegistrationForm.php
@@ -11,7 +11,7 @@ use App\Models\MealModel;
 use App\Models\MeetingModel;
 use Nette\Application\UI\Form;
 use Nette\Forms\Controls;
-use App\Services\UserService;
+use App\Services\SkautIS\UserService;
 
 class RegistrationForm extends BaseForm
 {
@@ -162,7 +162,7 @@ class RegistrationForm extends BaseForm
 				->addRule(Form::MAX_LENGTH, static::MESSAGE_MAX_LENGTH, 50);
 
 		$form->addSelect('province', 'Kraj:', $provinces)
-			->setPrompt('Zvolte kraj');
+			->setPrompt('zvolte kraj');
 
 		$form = $this->buildMealSwitcher($form);